requirements.txt
django==1.8.2
pymysql
django-tinymce
django-haystack==2.6
django-redis==4.4
django-celery
itsdangerous
Pillow
whoosh
AliPay==0.7.4
fdfs_client